// src/app/core/enums/info-results.enum.ts
export enum EInfoResults {
  getting = "Obteniendo datos",
  analyzing = "Analizando | No apuestes, se paciente.",
  changing_to_up = "Estableciendo Techo | Cambió la semilla",
  changing_to_down = "Estableciendo Piso | Cambió la semilla",
  prediction = "Predicción Establecida",
  frecuency_calculating = "Calculando Frecuencia",
  maybe_pink = "Posible Rosa muy cerca",
  chaging_trend = "Cambiando Tendencia"
}